raspbberrypi_send_to_db/requirements.txt
azure-eventhub
python-dotenv
azure-eventhub-checkpointstoreblob-aio
azure-cosmos